requirements.txt
pip
rich
pygame-ce
